# HttpObjectDecoder performance improvements