@Scottmitch 2015-07-29
Here are the benchmark results from the new benchmark introduced in this PR:

```
Benchmark                                                  (charsInitSize)  (simulatedDataSize)   Mode  Cnt          Score          Error  Units
AppendableCharSequenceBenchmark.appendCatchExceptionAfter               32                   10  thrpt   10  163214901.273 ±   827188.376  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter               32                  100  thrpt   10  174286555.792 ±  2652725.651  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter               32                10000  thrpt   10  172427704.206 ±  1050216.420  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter               32              1000000  thrpt   10  198293367.249 ±  2421629.155  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter               64                   10  thrpt   10  142145589.241 ±  8645429.915  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter               64                  100  thrpt   10  181533248.649 ±  5987405.075  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter               64                10000  thrpt   10  173277417.401 ±  1248384.171  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter               64              1000000  thrpt   10  189182569.469 ± 14148118.324  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter              128                   10  thrpt   10  115053210.485 ± 13577182.542  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter              128                  100  thrpt   10  223726027.774 ±  1324390.093  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter              128                10000  thrpt   10  165406416.797 ± 11147674.474  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter              128              1000000  thrpt   10  198220873.585 ±  1147101.825  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter              256                   10  thrpt   10   74111128.012 ±  1085650.658  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter              256                  100  thrpt   10  193344817.439 ± 15107053.293  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter              256                10000  thrpt   10  165729916.943 ± 11514247.799  ops/s
AppendableCharSequenceBenchmark.appendCatchExceptionAfter              256              1000000  thrpt   10  195141113.643 ±  5004763.175  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                   32                   10  thrpt   10  113326738.596 ±  3602148.520  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                   32                  100  thrpt   10  161737695.599 ±  2077902.839  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                   32                10000  thrpt   10  139809217.421 ± 16441381.540  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                   32              1000000  thrpt   10  167939658.809 ±  1512346.827  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                   64                   10  thrpt   10  119860349.224 ±  1988225.873  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                   64                  100  thrpt   10  146528435.155 ± 10373345.144  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                   64                10000  thrpt   10  136851264.438 ± 16170124.429  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                   64              1000000  thrpt   10  155422756.555 ± 18484402.612  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                  128                   10  thrpt   10  107271184.043 ±  8146907.839  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                  128                  100  thrpt   10  140162088.632 ± 59628360.025  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                  128                10000  thrpt   10  139645092.002 ± 15219642.375  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                  128              1000000  thrpt   10  165063516.815 ±  3175104.204  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                  256                   10  thrpt   10   69477367.568 ±   693150.372  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                  256                  100  thrpt   10  132924409.321 ± 46530260.424  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                  256                10000  thrpt   10  146548304.137 ±  6739326.894  ops/s
AppendableCharSequenceBenchmark.appendCheckBeforeCopy                  256              1000000  thrpt   10  164370777.452 ±  5629998.537  ops/s
```