@normanmaurer 2015-07-29
@Scottmitch could you check the test-failure ?