@Scottmitch 2015-07-29
@normanmaurer - Fixed.

Perf results from a `wrk` benchmark:

== This PR
```
  16 threads and 256 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    17.51ms   21.70ms 902.21ms   84.96%
    Req/Sec   162.74k    31.96k  290.59k    72.81%
  309461751 requests in 2.00m, 41.79GB read
Requests/sec: 2576716.01
Transfer/sec:    356.32MB
```

== Before this PR
```
Running 2m test @ http://17.177.98.59:8080/plaintext
  16 threads and 256 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    18.95ms   29.67ms   1.23s    87.39%
    Req/Sec   157.67k    33.29k  272.42k    72.60%
  299857163 requests in 2.00m, 40.49GB read
Requests/sec: 2496765.40
Transfer/sec:    345.26MB
```