@normanmaurer 2015-07-30
@Scottmitch please cherry-pick!