@Scottmitch 2015-07-30
Cherry-picked 4.0 (1fcc72a) 4.1 (a771306) master (9d347ff)